--- HW2/5.function_scope.py
@@ -1,10 +1,0 @@
-x = 100
-
-
-def change_var(x=None):
-    x = 5
-    print(x)
-
-
-change_var(x)
-print(x)
