+++ Lesson_2/5.function_scope.py
@@ -1,0 +1,11 @@
+x = 100
+
+
+def change_var():
+#    global x
+    x = 5
+    print(x)
+
+#
+change_var()
+print(x)
