+++ Lesson_2/7. decorator.py
